made remaining seeders and renamed files for consistency

+++ migrations/createTable-meet_greets.js
@@ -7,7 +7,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER
+        type: Sequelize.SMALLINT
       },
       event_id: {
         allowNull: false,
