+++ migrations/createTable-stage_events.js
@@ -7,7 +7,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER
+        type: Sequelize.SMALLINT
       },
       stage_id: {
         allowNull: false,
